支持群消息，简化代码

+++ config.py
@@ -1,34 +1,33 @@
 import asyncio
 from graia.broadcast import Broadcast
 from graia.application import GraiaMiraiApplication
-from graia.application.message.elements.internal import Plain
 from graia.application.session import Session
-from graia.application.message.chain import MessageChain
-from graia.application.friend import Friend
 
 import time
 import os
-import re
 
 
 def writeFile(id, suffix, code):
-    path = 'temp/' + str(id)
+    path = 'temp/' + id
     folder = os.path.exists(path)
     if not folder:
         os.makedirs(path)
     
-    now = time.strftime('%Y-%m-%d %H-%M-%S',time.localtime(time.time()))
-    fileName = 'temp/' + str(id) + '/' + str(now) + suffix
+    now = time.strftime('%Y-%m-%d %H-%M-%S',time.localtime(time.time()))     
+    fileName = path + '/' + str(now) + suffix
+    if os.path.exists(fileName):
+        raise NameError('Too Frequent')
+
     f = open(fileName, 'w', encoding='utf-8')
     f.write(code + "\n")
     f.close()
     return(fileName)
 
+
 def runCMD(cmd):
     result = os.popen(cmd)
-    ans = result.read()
+    ans = result.read().strip("\n")
     return(ans)
-
 
 loop = asyncio.get_event_loop()
 
@@ -45,17 +44,22 @@ app = GraiaMiraiApplication(
 
 
 rawHelp = [
-    [1, "ExecutePython (epy)", "运行 Python"],
-    [2, "ExecuteMathematica (ema)" , "运行 Mathematica"],
+    [1, "epy" , "ExecutePython"],
+    [2, "ema" , "ExecuteMathematica"],
     [3, "help" , "帮助"],
-    [4, "pip " , "Python 库安装"]
+    [4, "pip install" , "Python 库安装"]
 ]
 
 help = '\n'.join(['(' + str(l[0]) + ') ' + l[1] + '  : ' + l[2] for l in rawHelp])
+
 
 def regularQ(firstLine,full,kernel):
     if firstLine.find(kernel)<=0:
         return False
     left = firstLine.split(kernel)[0]
     right = firstLine[len(left):].split(' ')[0]
-    return("execute".startswith(left) and full.startswith(right))
+    return "execute".startswith(left) and full.startswith(right)
+
+
+if __name__=='__main__':
+    print(regularQ('epy','python','py'))
